+++ tests/core/packageVersion.test.ts
@@ -1,66 +1,5 @@
 import { describe, it, expect } from 'vitest';
-
-// We need to test the package parsing function, but it's not exported
-// Let's create a simplified version for testing
-function parsePackageFromCommand(command: string, args: string[] = []): { name: string; version?: string | undefined } | null {
-  const fullCommand = [command, ...args].join(' ');
-
-  // Pattern 1: npx package-name@version
-  const npxVersionMatch = fullCommand.match(/npx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*([^@\s]+)@([^\s]+)/);
-  if (npxVersionMatch && npxVersionMatch[1] && npxVersionMatch[2]) {
-    return { name: npxVersionMatch[1], version: npxVersionMatch[2] };
-  }
-
-  // Pattern 2: npx -y @scope/package@latest or npx @scope/package@latest
-  const npxScopedMatch = fullCommand.match(/npx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*(@[^/]+\/[^@\s]+)(?:@([^\s]+))?/);
-  if (npxScopedMatch && npxScopedMatch[1]) {
-    return { name: npxScopedMatch[1], version: npxScopedMatch[2] || undefined };
-  }
-
-  // Pattern 3: npx package-name (without version)
-  const npxMatch = fullCommand.match(/npx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*([^@\s]+)/);
-  if (npxMatch && npxMatch[1]) {
-    return { name: npxMatch[1] };
-  }
-
-  // Pattern 4: bunx package-name@version (same patterns as npx)
-  const bunxVersionMatch = fullCommand.match(/bunx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*([^@\s]+)@([^\s]+)/);
-  if (bunxVersionMatch && bunxVersionMatch[1] && bunxVersionMatch[2]) {
-    return { name: bunxVersionMatch[1], version: bunxVersionMatch[2] };
-  }
-
-  // Pattern 5: bunx -y @scope/package@latest or bunx @scope/package@latest
-  const bunxScopedMatch = fullCommand.match(/bunx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*(@[^/]+\/[^@\s]+)(?:@([^\s]+))?/);
-  if (bunxScopedMatch && bunxScopedMatch[1]) {
-    return { name: bunxScopedMatch[1], version: bunxScopedMatch[2] || undefined };
-  }
-
-  // Pattern 6: bunx package-name (without version)
-  const bunxMatch = fullCommand.match(/bunx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*([^@\s]+)/);
-  if (bunxMatch && bunxMatch[1]) {
-    return { name: bunxMatch[1] };
-  }
-
-  // Pattern 7: pipx run --spec package==version binary
-  const pipxSpecMatch = fullCommand.match(/pipx\s+run\s+--spec\s+([^=\s]+)==([^\s]+)/);
-  if (pipxSpecMatch && pipxSpecMatch[1] && pipxSpecMatch[2]) {
-    return { name: pipxSpecMatch[1], version: pipxSpecMatch[2] };
-  }
-
-  // Pattern 8: pipx run package-name (without version)
-  const pipxMatch = fullCommand.match(/pipx\s+run\s+(?:--python\s+[^\s]+\s+)?([^@\s]+)/);
-  if (pipxMatch && pipxMatch[1]) {
-    return { name: pipxMatch[1] };
-  }
-
-  // Pattern 9: uvx package-name or uvx --from git+... package-name
-  const uvxMatch = fullCommand.match(/uvx\s+(?:--from\s+[^\s]+\s+)?([^@\s]+)/);
-  if (uvxMatch && uvxMatch[1]) {
-    return { name: uvxMatch[1] };
-  }
-
-  return null;
-}
+import { parsePackageFromCommand } from '../../src/core/packageVersionUtils.js';
 
 describe('Package Version Detection', () => {
   describe('parsePackageFromCommand', () => {
