+++ tests/core/packageVersionIntegration.test.ts
@@ -1,150 +1,10 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { MCPServerType } from '../../src/types/index.js';
+import { getPackageVersion } from '../../src/core/packageVersionUtils.js';
 
 // Mock fetch globally
 const mockFetch = vi.fn();
 global.fetch = mockFetch;
-
-// We need to import the internal functions for testing
-// Since they're not exported, we'll create a test wrapper
-class PackageVersionTester {
-  static parsePackageFromCommand(command: string, args: string[] = []): { name: string; version?: string | undefined } | null {
-    const fullCommand = [command, ...args].join(' ');
-
-    // Pattern 1: npx package-name@version
-    const npxVersionMatch = fullCommand.match(/npx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*([^@\s]+)@([^\s]+)/);
-    if (npxVersionMatch && npxVersionMatch[1] && npxVersionMatch[2]) {
-      return { name: npxVersionMatch[1], version: npxVersionMatch[2] };
-    }
-
-    // Pattern 2: npx -y @scope/package@latest or npx @scope/package@latest
-    const npxScopedMatch = fullCommand.match(/npx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*(@[^/]+\/[^@\s]+)(?:@([^\s]+))?/);
-    if (npxScopedMatch && npxScopedMatch[1]) {
-      return { name: npxScopedMatch[1], version: npxScopedMatch[2] || undefined };
-    }
-
-    // Pattern 3: npx package-name (without version)
-    const npxMatch = fullCommand.match(/npx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*([^@\s]+)/);
-    if (npxMatch && npxMatch[1]) {
-      return { name: npxMatch[1] };
-    }
-
-    // Pattern 4: bunx package-name@version (same patterns as npx)
-    const bunxVersionMatch = fullCommand.match(/bunx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*([^@\s]+)@([^\s]+)/);
-    if (bunxVersionMatch && bunxVersionMatch[1] && bunxVersionMatch[2]) {
-      return { name: bunxVersionMatch[1], version: bunxVersionMatch[2] };
-    }
-
-    // Pattern 5: bunx -y @scope/package@latest or bunx @scope/package@latest
-    const bunxScopedMatch = fullCommand.match(/bunx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*(@[^/]+\/[^@\s]+)(?:@([^\s]+))?/);
-    if (bunxScopedMatch && bunxScopedMatch[1]) {
-      return { name: bunxScopedMatch[1], version: bunxScopedMatch[2] || undefined };
-    }
-
-    // Pattern 6: bunx package-name (without version)
-    const bunxMatch = fullCommand.match(/bunx\s+(?:-[ygc]\s+)?(?:--\S+\s+)*([^@\s]+)/);
-    if (bunxMatch && bunxMatch[1]) {
-      return { name: bunxMatch[1] };
-    }
-
-    // Pattern 7: pipx run --spec package==version binary
-    const pipxSpecMatch = fullCommand.match(/pipx\s+run\s+--spec\s+([^=\s]+)==([^\s]+)/);
-    if (pipxSpecMatch && pipxSpecMatch[1] && pipxSpecMatch[2]) {
-      return { name: pipxSpecMatch[1], version: pipxSpecMatch[2] };
-    }
-
-    // Pattern 8: pipx run package-name (without version)
-    const pipxMatch = fullCommand.match(/pipx\s+run\s+(?:--python\s+[^\s]+\s+)?([^@\s]+)/);
-    if (pipxMatch && pipxMatch[1]) {
-      return { name: pipxMatch[1] };
-    }
-
-    // Pattern 9: uvx package-name or uvx --from git+... package-name
-    const uvxMatch = fullCommand.match(/uvx\s+(?:--from\s+[^\s]+\s+)?([^@\s]+)/);
-    if (uvxMatch && uvxMatch[1]) {
-      return { name: uvxMatch[1] };
-    }
-
-    return null;
-  }
-
-  static async fetchPackageVersionFromNpm(packageName: string, timeoutMs = 5000): Promise<string | null> {
-    try {
-      const controller = new AbortController();
-      const timeoutId = setTimeout(() => controller.abort(), timeoutMs);
-
-      const url = `https://registry.npmjs.org/${encodeURIComponent(packageName)}`;
-      const response = await fetch(url, {
-        signal: controller.signal,
-        headers: { 'Accept': 'application/json' }
-      });
-
-      clearTimeout(timeoutId);
-
-      if (!response.ok) {
-        return null;
-      }
-
-      const data = await response.json();
-      return data['dist-tags']?.latest || data.version || null;
-    } catch (error) {
-      return null;
-    }
-  }
-
-  static async fetchPackageVersionFromPyPI(packageName: string, timeoutMs = 5000): Promise<string | null> {
-    try {
-      const controller = new AbortController();
-      const timeoutId = setTimeout(() => controller.abort(), timeoutMs);
-
-      const url = `https://pypi.org/pypi/${encodeURIComponent(packageName)}/json`;
-      const response = await fetch(url, {
-        signal: controller.signal,
-        headers: { 'Accept': 'application/json' }
-      });
-
-      clearTimeout(timeoutId);
-
-      if (!response.ok) {
-        return null;
-      }
-
-      const data = await response.json();
-      return data.info?.version || null;
-    } catch (error) {
-      return null;
-    }
-  }
-
-  static isPythonPackageManager(command: string, args: string[] = []): boolean {
-    const fullCommand = [command, ...args].join(' ');
-    return fullCommand.includes('pipx') || fullCommand.includes('uvx');
-  }
-
-  static async getPackageVersion(server: { type: string; command?: string; args?: string[] }): Promise<string> {
-    if (server.type !== MCPServerType.STDIO || !server.command) {
-      return "unknown";
-    }
-
-    const packageInfo = this.parsePackageFromCommand(server.command, server.args);
-    if (!packageInfo) {
-      return "unknown";
-    }
-
-    // If we already have a version from the command, use it
-    if (packageInfo.version && packageInfo.version !== 'latest') {
-      return packageInfo.version;
-    }
-
-    // Determine which registry to use based on package manager
-    const usePyPI = this.isPythonPackageManager(server.command, server.args);
-    const latestVersion = usePyPI
-      ? await this.fetchPackageVersionFromPyPI(packageInfo.name)
-      : await this.fetchPackageVersionFromNpm(packageInfo.name);
-
-    return latestVersion || "unknown";
-  }
-}
 
 describe('Package Version Detection (Integration)', () => {
   beforeEach(() => {
@@ -166,12 +26,13 @@ describe('Package Version Detection (Integration)', () => {
       });
 
       const server = {
+        name: 'test-server',
         type: MCPServerType.STDIO,
         command: 'npx',
         args: ['test-package']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
 
       expect(mockFetch).toHaveBeenCalledWith(
         'https://registry.npmjs.org/test-package',
@@ -191,12 +52,13 @@ describe('Package Version Detection (Integration)', () => {
       });
 
       const server = {
+        name: 'test-server',
         type: MCPServerType.STDIO,
         command: 'bunx',
         args: ['-y', 'another-package']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
 
       expect(mockFetch).toHaveBeenCalledWith(
         'https://registry.npmjs.org/another-package',
@@ -214,12 +76,13 @@ describe('Package Version Detection (Integration)', () => {
       });
 
       const server = {
+        name: 'test-server',
         type: MCPServerType.STDIO,
         command: 'npx',
         args: ['nonexistent-package']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
       expect(version).toBe('unknown');
     });
   });
@@ -234,12 +97,13 @@ describe('Package Version Detection (Integration)', () => {
       });
 
       const server = {
+        name: 'test-server',
         type: MCPServerType.STDIO,
         command: 'pipx',
         args: ['run', 'poetry']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
 
       expect(mockFetch).toHaveBeenCalledWith(
         'https://pypi.org/pypi/poetry/json',
@@ -259,12 +123,13 @@ describe('Package Version Detection (Integration)', () => {
       });
 
       const server = {
+        name: 'test-server',
         type: MCPServerType.STDIO,
         command: 'uvx',
         args: ['black']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
 
       expect(mockFetch).toHaveBeenCalledWith(
         'https://pypi.org/pypi/black/json',
@@ -279,12 +144,13 @@ describe('Package Version Detection (Integration)', () => {
       mockFetch.mockRejectedValueOnce(new Error('Network error'));
 
       const server = {
+        name: 'test-server',
         type: MCPServerType.STDIO,
         command: 'uvx',
         args: ['failing-package']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
       expect(version).toBe('unknown');
     });
   });
@@ -292,12 +158,13 @@ describe('Package Version Detection (Integration)', () => {
   describe('version extraction from commands', () => {
     it('should extract version directly from command without API calls', async () => {
       const server = {
+        name: 'test-server',
         type: MCPServerType.STDIO,
         command: 'npx',
         args: ['test-package@4.5.6']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
 
       expect(mockFetch).not.toHaveBeenCalled();
       expect(version).toBe('4.5.6');
@@ -305,12 +172,13 @@ describe('Package Version Detection (Integration)', () => {
 
     it('should extract version from pipx --spec commands', async () => {
       const server = {
+        name: 'test-server',
         type: MCPServerType.STDIO,
         command: 'pipx',
         args: ['run', '--spec', 'poetry==1.7.1', 'poetry']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
 
       expect(mockFetch).not.toHaveBeenCalled();
       expect(version).toBe('1.7.1');
@@ -320,12 +188,13 @@ describe('Package Version Detection (Integration)', () => {
   describe('non-package manager commands', () => {
     it('should return unknown for non-package-manager commands', async () => {
       const server = {
+        name: 'test-server',
         type: MCPServerType.STDIO,
         command: 'python',
         args: ['-m', 'some_module']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
 
       expect(mockFetch).not.toHaveBeenCalled();
       expect(version).toBe('unknown');
@@ -333,12 +202,13 @@ describe('Package Version Detection (Integration)', () => {
 
     it('should return unknown for non-STDIO servers', async () => {
       const server = {
-        type: 'HTTP',
+        name: 'test-server',
+        type: 'HTTP' as any,
         command: 'npx',
         args: ['test-package']
       };
 
-      const version = await PackageVersionTester.getPackageVersion(server);
+      const version = await getPackageVersion(server);
 
       expect(mockFetch).not.toHaveBeenCalled();
       expect(version).toBe('unknown');
